+++ packages/opencensus-core/src/trace/model/span-base.ts
@@ -18,6 +18,9 @@ import {Clock} from '../../internal/clock';
 import {randomSpanId} from '../../internal/util';
 import * as types from './types';
 
+const STATUS_OK = {
+  code: types.CanonicalCode.OK
+};
 
 /** Defines a base model for spans. */
 export abstract class SpanBase implements types.Span {
@@ -52,7 +55,7 @@ export abstract class SpanBase implements types.Span {
   /** Kind of span. */
   kind: string = null;
   /** A final status for this span */
-  status: number;
+  status: types.Status = STATUS_OK;
   /** set isRootSpan  */
   abstract get isRootSpan(): boolean;
 
@@ -181,6 +184,15 @@ export abstract class SpanBase implements types.Span {
       'id': id,
       'timestamp': timestamp ? timestamp : Date.now(),
     } as types.MessageEvent);
+  }
+
+  /**
+   * Sets a status to the span.
+   * @param code The canonical status code.
+   * @param message optional A developer-facing error message.
+   */
+  setStatus(code: types.CanonicalCode, message?: string) {
+    this.status = {code, message};
   }
 
   /** Starts the span. */