+++ packages/opencensus-instrumentation-http2/src/http2.ts
@@ -141,8 +141,8 @@ export class Http2Plugin extends HttpPlugin {
         span.addAttribute(
             Http2Plugin.ATTRIBUTE_HTTP_STATUS_CODE,
             `${responseHeaders[':status']}`);
-        span.status =
-            Http2Plugin.convertTraceStatus(+responseHeaders[':status']);
+        span.setStatus(
+            Http2Plugin.parseResponseStatus(+responseHeaders[':status']));
       });
 
       request.on('end', () => {
@@ -258,7 +258,7 @@ export class Http2Plugin extends HttpPlugin {
                 Http2Plugin.ATTRIBUTE_HTTP_USER_AGENT, userAgent);
             rootSpan.addAttribute(
                 Http2Plugin.ATTRIBUTE_HTTP_STATUS_CODE, `${statusCode}`);
-            rootSpan.status = Http2Plugin.convertTraceStatus(statusCode);
+            rootSpan.setStatus(Http2Plugin.parseResponseStatus(statusCode));
 
             rootSpan.addMessageEvent(
                 'MessageEventTypeRecv', uuid.v4().split('-').join(''));